Updated var types, prototypes and more

+++ src/print/ft_putpointer_fd.c
@@ -6,7 +6,7 @@
 /*   By: ppontet <ppontet@student.42lyon.fr>        +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/11/20 21:10:56 by ppontet           #+#    #+#             */
-/*   Updated: 2025/02/17 18:24:12 by ppontet          ###   ########lyon.fr   */
+/*   Updated: 2025/03/21 12:26:45 by ppontet          ###   ########lyon.fr   */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -14,7 +14,7 @@
 
 /**
  * @brief Print pointer adress
- * 
+ *
  * @param ptr pointer to print
  * @return ssize_t number of char printed
  */
@@ -23,5 +23,5 @@ ssize_t	ft_putpointer_fd(void *ptr)
 	if (!ptr)
 		return (write(1, "(nil)", 5));
 	write(1, "0x", 2);
-	return (2 + ft_putnbr_hex((unsigned long)ptr, 'x'));
+	return (2 + ft_putnbr_hex(((int)(unsigned long)ptr), 'x'));
 }
